Get system tests running on datastore emulator
Apparently upstream system tests don't work on the emulator.  This seems
to be just from not being maintained, so this commit fixes them.  It's
unclear to me if this is desired upstream, but it's here for testing
purposes if you need it!

Test plan:
```
gcloud beta emulators datastore start --project testytest
DATASTORE_EMULATOR_HOST=localhost:8081 DATASTORE_DATASET=testytest \
  GOOGLE_APPLICATION_CREDENTIALS=/dev/null \
  nox -s system-3.7 -- --tb=native
```